+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 - Do not mistake "océan indien" as IN
 - Support for Saint-Martin (France) postcodes
+- Use French region names when finding French subdivisions
+  (the dept of the regional prefecture is used if a region name matched)
 
 ### Dependencies