+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [5.2.1] - 2024-05-02
+
+- "Southern Africa" no longer matches South Africa.
+
 ## [5.2.0] - 2024-04-08
 
 - Use French region names when finding French subdivisions